refactor(ui): extract InterfaceManager into modular components
Reduce InterfaceManager complexity by extracting responsibilities:
- DragDropCoordinator: All drag-drop handling (tab + @dnd-kit)
- WindowRenderer: Window rendering and layout logic
- EditModeUI: Edit mode overlay and indicators
- DndKitOverlay: @dnd-kit drag overlay component
- InterfaceManager now serves as orchestration layer
- Update exports and types accordingly
